+++ packages/eyes-selenium/src/selenium3/SpecWrappedDriver.js
@@ -1,6 +1,8 @@
+const {TypeUtils} = require('@applitools/eyes-sdk-core')
+const {By, Builder, until} = require('selenium-webdriver')
+const cmd = require('selenium-webdriver/lib/command')
 const SeleniumFrame = require('../SeleniumFrame')
 const SeleniumWrappedElement = require('../SeleniumWrappedElement')
-const cmd = require('selenium-webdriver/lib/command')
 
 module.exports = {
   isEqualFrames(leftFrame, rightFrame) {
@@ -39,6 +41,9 @@ module.exports = {
   },
   async findElement(driver, selector) {
     try {
+      if (TypeUtils.isString(selector)) {
+        selector = By.css(selector)
+      }
       return await driver.findElement(selector)
     } catch (err) {
       if (err.name === 'NoSuchElementError') return null
@@ -46,6 +51,9 @@ module.exports = {
     }
   },
   async findElements(driver, selector) {
+    if (TypeUtils.isString(selector)) {
+      selector = By.css(selector)
+    }
     return driver.findElements(selector)
   },
   async getWindowLocation(driver) {
@@ -62,11 +70,18 @@ module.exports = {
       .setPosition(location.x, location.y)
   },
   async getWindowSize(driver) {
-    const {width, height} = await driver
-      .manage()
-      .window()
-      .getSize()
-    return {width, height}
+    try {
+      const {width, height} = await driver
+        .manage()
+        .window()
+        .getSize()
+      return {width, height}
+    } catch (err) {
+      // workaround for Appium
+      return driver.execute(
+        new cmd.Command(cmd.Name.GET_WINDOW_SIZE).setParameter('windowHandle', 'current'),
+      )
+    }
   },
   async setWindowSize(driver, size) {
     await driver
@@ -129,5 +144,36 @@ module.exports = {
   },
   async visit(driver, url) {
     return driver.get(url)
+  },
+
+  /********* for testing purposes */
+  async build({capabilities, serverUrl = process.env.CVG_TESTS_REMOTE}) {
+    if (capabilities['sauce:options']) {
+      capabilities.username = capabilities['sauce:options'].username
+      capabilities.accesskey = capabilities['sauce:options'].accesskey
+    }
+    return new Builder()
+      .withCapabilities(capabilities)
+      .usingServer(serverUrl)
+      .build()
+  },
+
+  async cleanup(driver) {
+    return driver.quit()
+  },
+
+  async click(_driver, el) {
+    return el.click()
+  },
+
+  async waitUntilDisplayed(driver, selector, timeout) {
+    const el = await this.findElement(driver, selector)
+    return driver.wait(until.elementIsVisible(el), timeout)
+  },
+
+  async getElementRect(_driver, element) {
+    const location = await element.getLocation()
+    const size = await element.getSize()
+    return {...size, ...location}
   },
 }