+++ packages/sdk-coverage-tests/src/index.js
@@ -1,6 +1,4 @@
 const {makeEmitTracker} = require('./code-export')
 const {makeCoverageTests} = require('./tests')
-const TestSetup = require('../coverage-tests/util/TestSetup')
-const ApiAssertions = require('../coverage-tests/util/ApiAssertions')
 
-module.exports = {makeEmitTracker, makeCoverageTests, TestSetup, ApiAssertions}
+module.exports = {makeEmitTracker, makeCoverageTests}